Add basic models

+++ db/migrate/20220912171034_add_details_to_bills.rb
@@ -1,6 +1,6 @@
 class AddDetailsToBills < ActiveRecord::Migration[7.0]
   def change
     add_column :bills, :name, :string
-    add_column :bills, :amount, :decimal
+    add_column :bills, :amount, :integer
   end
 end
